Determine relationship type for Edge as part of Graph::build()
